pyproject.toml
[tool.poetry]
name = "iot-dashboard"
version = "0.1.0"
description = ""
authors = ["Tom Nordal <tnordal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
paho-mqtt = "^2.1.0"
python-dotenv = "^1.0.1"
dash = "^2.18.2"
plotly = "^5.24.1"
influxdb-client = "^1.48.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
